requirements.txt
# Core Data Engineering Libraries
pyspark==3.5.0
delta-spark==3.0.0

# Kafka
kafka-python==2.0.2
confluent-kafka==2.3.0

# Database
psycopg2-binary
sqlalchemy==2.0.23

# Airflow
apache-airflow==2.7.1
apache-airflow-providers-postgres==5.7.1
apache-airflow-providers-apache-spark==4.3.0
apache-airflow-providers-apache-kafka==1.10.0

# Data Quality
great-expectations==0.17.23

# Data Processing
pandas
numpy
pyarrow

# Configuration and Utilities
pyyaml==6.0.1
python-dotenv==1.0.0
click==8.1.7

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0

# Monitoring and Logging
structlog==23.2.0
prometheus-client==0.19.0

# Jupyter for Analysis
jupyter==1.0.0
ipykernel==6.27.1
matplotlib==3.8.2
seaborn==0.13.0

# Additional Utilities
requests==2.31.0
faker==20.1.0
python-dateutil==2.8.2
pytz==2023.3 